+++ pkg/collector/corechecks/networkpath/networkpath.go
@@ -19,6 +19,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/aggregator/sender"
 	"github.com/DataDog/datadog-agent/pkg/collector/check"
 	core "github.com/DataDog/datadog-agent/pkg/collector/corechecks"
+	"github.com/DataDog/datadog-agent/pkg/networkpath/payload"
 	"github.com/DataDog/datadog-agent/pkg/util/hostname"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/DataDog/datadog-agent/pkg/util/optional"
@@ -110,7 +111,7 @@ func (c *Check) getCommonTagsForMetrics() []string {
 }
 
 // SendNetPathMDToEP sends a traced network path to EP
-func (c *Check) SendNetPathMDToEP(sender sender.Sender, path traceroute.NetworkPath) error {
+func (c *Check) SendNetPathMDToEP(sender sender.Sender, path payload.NetworkPath) error {
 	payloadBytes, err := json.Marshal(path)
 	if err != nil {
 		return fmt.Errorf("error marshalling device metadata: %s", err)
@@ -120,7 +121,7 @@ func (c *Check) SendNetPathMDToEP(sender sender.Sender, path traceroute.NetworkP
 	return nil
 }
 
-func (c *Check) submitTelemetryMetrics(senderInstance sender.Sender, path traceroute.NetworkPath, startTime time.Time, tags []string) {
+func (c *Check) submitTelemetryMetrics(senderInstance sender.Sender, path payload.NetworkPath, startTime time.Time, tags []string) {
 	newTags := utils.CopyStrings(tags)
 
 	checkDuration := time.Since(startTime)
