+++ pkg/networkpath/payload/payload.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2024-present Datadog, Inc.
 
-package traceroute
+package payload
 
 import "github.com/DataDog/datadog-agent/pkg/network"
