+++ pkg/networkpath/traceroute/traceroute_linux.go
@@ -12,6 +12,7 @@ import (
 	"encoding/json"
 
 	dd_config "github.com/DataDog/datadog-agent/pkg/config"
+	"github.com/DataDog/datadog-agent/pkg/networkpath/payload"
 	"github.com/DataDog/datadog-agent/pkg/process/net"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
@@ -36,22 +37,22 @@ func New(cfg Config) (*LinuxTraceroute, error) {
 }
 
 // Run executes a traceroute
-func (l *LinuxTraceroute) Run(_ context.Context) (NetworkPath, error) {
+func (l *LinuxTraceroute) Run(_ context.Context) (payload.NetworkPath, error) {
 	tu, err := net.GetRemoteSystemProbeUtil(
 		dd_config.SystemProbe.GetString("system_probe_config.sysprobe_socket"))
 	if err != nil {
 		log.Warnf("could not initialize system-probe connection: %s", err.Error())
-		return NetworkPath{}, err
+		return payload.NetworkPath{}, err
 	}
 
 	resp, err := tu.GetTraceroute(clientID, l.cfg.DestHostname, l.cfg.DestPort, l.cfg.MaxTTL, l.cfg.TimeoutMs)
 	if err != nil {
-		return NetworkPath{}, err
+		return payload.NetworkPath{}, err
 	}
 
-	var path NetworkPath
+	var path payload.NetworkPath
 	if err := json.Unmarshal(resp, &path); err != nil {
-		return NetworkPath{}, err
+		return payload.NetworkPath{}, err
 	}
 
 	return path, nil
