+++ pkg/networkpath/traceroute/traceroute_windows.go
@@ -12,6 +12,7 @@ import (
 	"encoding/json"
 
 	dd_config "github.com/DataDog/datadog-agent/pkg/config"
+	"github.com/DataDog/datadog-agent/pkg/networkpath/payload"
 	"github.com/DataDog/datadog-agent/pkg/process/net"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
@@ -36,21 +37,21 @@ func New(cfg Config) (*WindowsTraceroute, error) {
 }
 
 // Run executes a traceroute
-func (w *WindowsTraceroute) Run(_ context.Context) (NetworkPath, error) {
+func (w *WindowsTraceroute) Run(_ context.Context) (payload.NetworkPath, error) {
 	tu, err := net.GetRemoteSystemProbeUtil(
 		dd_config.SystemProbe.GetString("system_probe_config.sysprobe_socket"))
 	if err != nil {
 		log.Warnf("could not initialize system-probe connection: %s", err.Error())
-		return NetworkPath{}, err
+		return payload.NetworkPath{}, err
 	}
 	resp, err := tu.GetTraceroute(clientID, w.cfg.DestHostname, w.cfg.DestPort, w.cfg.MaxTTL, w.cfg.TimeoutMs)
 	if err != nil {
-		return NetworkPath{}, err
+		return payload.NetworkPath{}, err
 	}
 
-	var path NetworkPath
+	var path payload.NetworkPath
 	if err := json.Unmarshal(resp, &path); err != nil {
-		return NetworkPath{}, err
+		return payload.NetworkPath{}, err
 	}
 
 	return path, nil
